requirements.txt
openai-whisper
gradio
transformers
torch
spacy
pydub
ffmpeg